[STEP 13, 14] 대기열 기능 개선

+++ src/main/java/kr/hhplus/be/server/api/reservation/domain/entity/Reservation.java
@@ -21,13 +21,13 @@ public class Reservation {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Long id;
 
-	private long userId;
+	private Long userId;
 
-	private long seatId;
+	private Long seatId;
 
-	private long seatNo;
+	private int seatNo;
 
 	private LocalDate concertDate;
